+++ packages/common/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 9.0.0
+
+### Breaking changes
+
+- Renamed `AccountTransactionType.TransferWithScheduleWithMemo` to `AccountTransactionType.TransferWithScheduleAndMemo`.
+
+### Changed
+
+- Bumped @concordium/rust-bindings to 1.1.0 (adds `display_type_schema_template` function)
+
+### Added
+
+- `getTransactionKindString` function.
+- `displayTypeSchemaTemplate` function.
+
 ## 8.0.0
 
 ### Breaking changes