+++ packages/web/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 6.0.0
+
+### Breaking changes
+
+- Bumped @concordium/rust-bindings to 1.1.0 and @concordium/common-sdk to 9.0.0. (adds `displayTypeSchemaTemplate/getTransactionKindString` and renames `AccountTransactionType.TransferWithScheduleWithMemo`)
+
 ## 5.0.0
 
 ### Breaking changes